Fixes for new MAC environment. AND On clients that do not initialize MPI themselves, do not monitor XIOS initialization with timers (requires that MPI_init has already been done)
git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS3/trunk@2471 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/attribute_map.cpp
@@ -401,7 +401,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -451,7 +451,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -501,7 +501,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -551,7 +551,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -586,7 +586,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -613,7 +613,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -648,7 +648,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -675,7 +675,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -710,7 +710,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -737,7 +737,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -775,7 +775,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -802,7 +802,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -840,7 +840,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -867,7 +867,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -905,7 +905,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
